+++ curlylint/lint.py
@@ -1,38 +1,14 @@
-from pathlib import Path
-
 import parsy
 
 from .check import check_files
 from .file import File
 from .issue import Issue, IssueLocation
 from .parse import make_parser
-from .util import flatten
 
 
 def get_parsy_error_location(error, file_path):
     line, column = parsy.line_info_at(error.stream, error.index)
     return IssueLocation(line=line, column=column, file_path=file_path)
-
-
-def resolve_file_paths_(input_name, extensions):
-    path = Path(input_name)
-    if not path.exists():
-        raise Exception("{} does not exist".format(path))
-
-    if path.is_dir():
-        return flatten(
-            resolve_file_paths_(child, extensions) for child in path.iterdir()
-        )
-
-    if not path.is_file():
-        raise Exception("{} is not a regular file".format(path))
-
-    return [path] if path.suffix in extensions else []
-
-
-def resolve_file_paths(input_names, extensions):
-    path_lists = (resolve_file_paths_(i, extensions) for i in input_names)
-    return flatten(path_lists)
 
 
 def parse_file(path_and_config):